+++ CandleEngine/src/Candle/ECS/BlueprintManager.h
@@ -11,11 +11,11 @@ namespace Candle {
 		public:
 			static Blueprint& Add(const std::string & name);
 			static void Add(Shared<Blueprint> blueprint);
-			static void Remove(long blueprintID);
+			static void Remove(size_t blueprintID);
 			static void Update();
 			static void Clear() { _blueprints.clear(); }
 
-			static std::unordered_map<long, Shared<Blueprint>> & All() { return _blueprints; }
+			static std::unordered_map<size_t, Shared<Blueprint>> & All() { return _blueprints; }
 			
 			template<typename T>
 			static std::vector<Shared<Blueprint>> & Get()
@@ -32,7 +32,7 @@ namespace Candle {
 			}
 
 		private:
-			static std::unordered_map<long, Shared<Blueprint>> _blueprints;
+			static std::unordered_map<size_t, Shared<Blueprint>> _blueprints;
 			static std::vector<Shared<Blueprint>> _batchBlueprints;
 	};
