+++ CandleEngine/src/Candle/ECS/ECS.h
@@ -41,10 +41,10 @@ namespace Candle {
 			static Blueprint& New(const std::string & name = "candleBlueprint");
 			static void Add(Shared<Blueprint> blueprint);
 			static void Add(System * system);
-			static void Remove(uint32_t blueprintID);
+			static void Remove(size_t blueprintID);
 			static void ClearBlueprints();
 
-			static Shared<Blueprint> GetBlueprint(long blueprintID);
+			static Shared<Blueprint> GetBlueprint(size_t blueprintID);
 
 			static void Update();
 			static void Render();
