+++ CandleEngine/src/Candle/Renderer/CameraManagement.h
@@ -14,7 +14,7 @@ namespace Candle {
 			static void Init();
 			static void UpdateView();
 			static void UseView(const glm::mat4& newView) { _viewProjectionMatrix = newView; }
-			static void RegisterMainCamera(long mainCameraBlueprintID);
+			static void RegisterMainCamera(size_t mainCameraBlueprintID);
 			static void Reset() { _viewProjectionMatrix = glm::mat4(0); }
 
 			static CameraManagement& Get() { return _instance; }
@@ -27,7 +27,7 @@ namespace Candle {
 			CameraManagement() {}
 
 			static CameraManagement _instance;
-			static long _mainCameraBlueprintID;
+			static size_t _mainCameraBlueprintID;
 			static glm::mat4 _viewProjectionMatrix;
 			static glm::mat4 _dummy;
